requirements.txt
python-dotenv>=1.0.1
tqdm>=4.66.1
requests>=2.31.0
numpy>=1.26.4

pypdf>=4.3.1

langchain>=0.2.14
langchain-core>=0.2.14
langchain-community>=0.4.1
langchain-openai>=0.1.23
langchain-text-splitters>=0.2.0

pinecone>=4.1.0
langchain-pinecone>=0.1.3

openai>=1.40.0

streamlit>=1.37.0
